+++ src/state_migration/tests/go-test/gen.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	gen "github.com/whyrusleeping/cbor-gen"
+)
+
+func main() {
+	if err := gen.WriteTupleEncodersToFile("./cbor_gen.go", "main",
+		StateRoot{},
+	); err != nil {
+		panic(err)
+	}
+}