+++ CHANGELOG.md
@@ -50,10 +50,16 @@
 
 - [#3126](https://github.com/ChainSafe/forest/issues/3126): Bail on database
   lookup errors instead of silently ignoring them.
+- [#2999](https://github.com/ChainSafe/forest/issues/2999): Restored `--tipset`
+  flag to `forest-cli snapshot export` to allow export at a specific tipset.
 
 ### Removed
 
 ### Fixed
+
+- [#3248](https://github.com/ChainSafe/forest/issues/3248): Fixed Forest being
+  unable to re-create its libp2p keypair from file and always changing its
+  `PeerId`.
 
 ## Forest v0.11.1 "Dagny Taggart"